Move prom metrics definition to a separate file
